chall_1/requirements.txt
Flask==2.3.2
Pillow==8.4.0
